Salesforce 2021 changes
Upgrade to icu4j 69.1
Add in Greenlandic/Kalaallisut
Some performance changes to reuse label sets in fallback languages (set useTranslatedLanguage=true in properties)
Refactoring of LanguageDeclension in an interface to support forwarding declensions
Allow overriding of the calendar in the locale provider to handle Arabic Calendar issues
TextDirection fixes to support script locale modifiers
Add in generators for JSLabels and Names.Dtd

+++ src/test/java/com/force/i18n/grammar/parser/GrammaticusLabelPerformanceTest.java
@@ -6,11 +6,14 @@
  */
 package com.force.i18n.grammar.parser;
 import java.lang.reflect.Field;
-import java.util.*;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 import java.util.logging.Logger;
 
-import com.force.i18n.*;
-import com.force.i18n.grammar.parser.GrammaticalLabelSetLoader;
+import com.force.i18n.HumanLanguage;
+import com.force.i18n.LanguageConstants;
+import com.force.i18n.LanguageProviderFactory;
 import com.google.common.collect.ImmutableList;
 
 public class GrammaticusLabelPerformanceTest extends BaseGrammaticalLabelTest {
@@ -69,7 +72,7 @@ public class GrammaticusLabelPerformanceTest extends BaseGrammaticalLabelTest {
      * @return
      */
     private double runCollisionCheckOnHash(Map<?,?> map) {
-        double m = (double) map.size();
+        double m = map.size();
         Set<Integer> s = new HashSet<Integer>(map.size() * 3);
         for (Object o : map.keySet()) {
             if (o != null)
